refactor : change package directory
