+++ src/test/java/step1/CalculatorTest.java
@@ -1,4 +1,4 @@
-package domain;
+package step1;
 
 import static org.assertj.core.api.Assertions.*;
 import static org.junit.jupiter.api.Assertions.*;
