+++ src/test/java/step1/SubtractionTest.java
@@ -1,4 +1,4 @@
-package domain;
+package step1;
 
 import static org.junit.jupiter.api.Assertions.*;
 
@@ -10,15 +10,15 @@ import org.junit.jupiter.api.Test;
 import step1.domain.Calculator;
 import step1.domain.Number;
 
-public class MultiplicationTest {
+public class SubtractionTest {
     @Test
-    @DisplayName("곱셈 연산")
-    public void multiplyTest() {
-        String operator = "*";
+    @DisplayName("뺄셈 연산")
+    public void subtractionTest() {
+        String operator = "-";
         Number left = Number.of(2);
         Number right = Number.of(3);
         Calculator calculator = new Calculator();
         Number result = calculator.calculate(left, right, operator);
-        assertEquals(result.getValue(), BigDecimal.valueOf(6));
+        assertEquals(BigDecimal.valueOf(-1), result.getValue());
     }
 }
